# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com
  description: This is a sample Gin server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: My Gin API
  version: "1.0"
paths:
  /api/v1/admin/logs:
    get:
      consumes:
      - application/json
      description: Get recent application logs (admin only)
      parameters:
      - default: 100
        description: Number of log lines to return
        in: query
        name: lines
        type: integer
      - description: Log level filter (error, warn, info, debug)
        in: query
        name: level
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Application logs
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get application logs
      tags:
      - admin
  /api/v1/admin/seed:
    post:
      consumes:
      - application/json
      description: Seed the database with initial users, questions, and exams (admin
        only)
      produces:
      - application/json
      responses:
        "200":
          description: Database seeded successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Seed database
      tags:
      - admin
...
swagger: "2.0"